requirements.txt
Django==3.1.2
django-ckeditor==6.0.0
django-crispy-forms==1.11.0
django-extensions==3.1.0
django-filter==2.4.0
django-js-asset==1.2.2
djangorestframework==3.12.2
python-dotenv==0.15.0
